[FIX] social: fix the note on the social post which is posted many time
Bug
===
When posting on multiple social account, a not is created in the chatter
of the social post. This note is posted many time.

If a live post fail, we can try to re-send it. In that case the note
will also be posted one per account selected on the social post.

Since 909e27b67a5281fbfe4a16a813fe8074823693b6 , we create a note even
if some live posts are not posted. So, we do not wait as before for the
post to be completed to create the note.

Solution
========
Log the note after the sending of all social live post, so we create one
and only one note in the chatter.

Task-2291254

X-original-commit ae919d70f8132bab531cb1ac47abe452ad688776

X-original-commit: 90770cf876273550b423c0c187ebd16323269fd0