[IMP] social_twitter: allow to ignore a mention when removing them
Purpose
=======
Allow to ignore a mention when we remove the mentions in a Tweet.

This commit is done in preparation of "Allowing social users to answer
to the tweets". We will need to allow the social user to mention the
author of the tweet to be able to answer to him.

Task-2293211

Part-of: odoo/enterprise#12173