[IMP] social_twitter: mention user when answering to a tweet
Purpose
=======
Be able to mention someone if we answer to one of his tweet (post/comment).

Before, all mentions were removed, but now, all mention will be removed
in the comment except the mention of the author of the post/comment which
we are answering.

Automatically add a mention (if not already present) when answering.
So, this new behavior will fix the following issue.

Bug fixed
=========
In social Twitter, in the feed view, we can see the tweets that reply
to our Tweet. When we reply to a Tweeter account (which is not our account),
the Tweet will not display as a response on Twitter.

1. Take 2 Twitter accounts (account A and B)
2. Link the account A in Odoo
3. With the account A, write a post (tweet)
4. With the account B, go on tweeter.com and add a response below the first tweet
5. Go back to Odoo, in the feed view, and with the account A open the first "post" and in the "comments" replies to the tweet of the account B
6. Go on Twitter with the account A, the last response will not be displayed as a response to the tweet of the account B
=> Solution: mention the user in the answer

Task-2293211

Part-of: odoo/enterprise#12173