[IMP] stock_barcode_picking_batch: cluster picking
Adds a new selection field on `stock.quant.package`, `package_use`.
On 'disposable', the package's use is still the same than before.
On 'reusable', the user will be able to scan the package's barcode when
processing a picking batch in the Barcode App.

The purpose of those packages is to be able to put your picks into one
or another package depending of the picking.
Also, when a barcode line has a `result_package_id` set, the app will
suggest to reuse this package for the other lines who share the same picking.

task-1884520
Related Community PR: odoo/odoo#55096
Upgrade PR: odoo/upgrade#1569
Closes: odoo/enterprise#12086