[FIX] hr_contract_salary: Fix submit of contract if married in marital status
Currently, In salary configurator form when we select married in the
marital status box then form can not be submited.

since the commit https://github.com/odoo/enterprise/commit/7945b6a252231cc357b9aada9ecc85bda418b1fb
Required is applied properly on spouse and other input but
when there is input type checkbox then there is only two possible
values 'checked/unchecked' so even if we uncheck the checkbox
is should not raise required error.
also it should not raise the required error when input without name.
as some dropdown input is without the name.

So in this commit, Fix the issue and do not raise reauired error
for input type checkbox and input without name.

closes odoo/enterprise#13033

Taskid: 2314164
Closes: #12445
X-original-commit: 34191c830cfcc40edeb275dfd7a17385d3ab6986
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>