[FIX] l10n_ar_edi: credit note with full reconcile
Formerly if we try to generate a credit note from a electronic invoice
will create and validate an electronic credit note. However if there an
error in AFIP validating the credit note we make a rollback and the
invoice is deletes, showing to the user error "Record does not exist or has
been deleted".

To avoid this, we ensure that the credit note is not created if there
was an error and show the proper information to the user about the
error we found out in AFIP. This way the user will be able to make the
proper fixes and then create/validate the credit note in both odoo and AFIP

closes odoo/enterprise#17189

X-original-commit: 1d284668f3f66fc37cdd01543134c7248ec0b93b
Signed-off-by: Josse Colpaert <jco@openerp.com>