[FIX] mrp_account_enterprise: fixes cost analysis report
Bunch of fixes for the cost analysis report of MO:
- Remove extra quantity feature, depends of the
state of the BoM when the MO was "Mark as Done".
If the BoM is changed after, it will be wrong.
- Change the link to the bom line, redirect to the product form instead
(same issue than above).
- Now, a WO is not always related to a operation.
Show WO name instead of operation name and
it will be a link only if related to a operation.
- Fix wrong quantity of raw material in case of change
after the mark as done.

task-40259

closes odoo/enterprise#12517

Related: odoo/odoo#56127
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>