[FIX] web_studio: don't add abstract model one2many
When adding a field in studio, it was possible to add a one2many field
of an abstract model, this caused access error.

With this changeset, one2many fields of abstract models will not be
shown when adding field with studio (this is already done for many2many
and many2one).

opw-2315371
closes #12426

closes odoo/enterprise#12435

X-original-commit: 553fb3443a7ceca19dac18902ec69d6f1570f5da
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>