[IMP] web_cohort: add invisible attribute for measures
PURPOSE

Be able to remove some measures from cohort view. Currently fields are
automatically added but some of them make no sense.

SPECIFICATIONS

Like done in community for pivot and graph views we want to be able to hide
some fields in measures dropdown. This commit adds the grammar to be able
to add child fields to the cohort view with the invisible attribute.

As we want to limit rng modification this grammar is done locally in web
corhort rng file.

We also take this opportunity to support the string attribute in order to be
able to update the string used as label for a given field.

LINKS

Task ID-2288381
Community PR #54774
Enterprise PR odoo/enterprise#11970
Upgrade PR odoo/upgrade#1519