[FIX] quality: no column shown in quality alert kanban
Currently no colunm will be shown in quality alert kanban view, no
matter we have quality alert stages or not. This is because in the
group_expand method of stage_id, no record will be found when provide no
information.

In this commit, we change _read_group_stage_ids method, when not enough
infomation is provided, we search stage with no team_ids.

Task 2310494
PR #12414

closes odoo/enterprise#13067

X-original-commit: bb5e8fe9b762ffc96985b19caf1dd91530c50acb
Related: odoo/odoo#57344
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>