[FIX] documents: Fix the view of archive/restore button of documents
Before this commit, the archive and restore buttons are displayed ugly
due the name at bottom of button. while in document inspector there is
no name in any button, only icons.

So in this commit, fix the display issue and do not dipslay the name
of the button at the bottom of button. only display the icons.

closes odoo/enterprise#12464

Taskid: 2320257
Closes: #12464
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>