[IMP] social_twitter: do not allow social users to spam
Purpose
=======
Do not allow the social user to spam Twitter user in the feed view.

If social users use Odoo to spam Twitter users, our Twitter application
credential can be restricted to the "Read-only mode". We want to avoid
this situation so we do not allow spam.

Task-2293211

Part-of: odoo/enterprise#12173