[FIX] project_forecast: Display all selected task with 'View Planning'
Purpose
=======

The commit 87fc7bc introduced a new action to display all the planning
for the selected tasks.

It was intended to display all the tasks, even if no planning.slot
is defined for them.

However, the _read_group_task_id used as a group_expand attribute
was already existing below in the code, an thus is overwritten.

This commit reintroduce the feature, by using a specific context key
for this use case.