[FIX] documents_spreadsheet: Fetch subgroup display names only once
If a pivot subgroup is a relational field (e.g. grouped by date:month > salesteam)
the display names of the relational records are fetched once
per parent group, resulting in many useless RPC calls.

This happens because all header evaluations are done asynchronously and they
all try to load the display name, even if it's already being fetched (but
not resolved yet).

Instead of waiting for the label then give it to the cache, the promise
is directly given to the cache. That way, the promise can be awaited by
any others.

closes odoo/enterprise#12502

Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>