[IMP] sale_amazon: use the token authentication with the Odoo proxy OoB
Forward-port to master (13.5) of odoo/enterprise#11289

This commit merges the changes of the 'stable'-released
`sale_amazon_authentication` module into the main `sale_amazon` module.

Since Amazon requested that we stop asking our customers to get their
own developper keys and instead use a token with our developper keys
for authentication, a proxy was deployed on one of our server to sign
all requests to the Amazon API with our keys and forward the response
back to the customer's instance.

In short, this commit:
- removes the developper keys fields from the amazon.account model and
  adds a `auth_token` field
- patches the mws library we used to contact our proxy instead of the
  Amazon API endpoints

Some smaller changes are included in this commit, since it is a squashed
version of the initial PR which included smaller fixes such as:
- support for the amazon.nl marketplace
- ensure a customer rank of 1 is set on partners created on the fly so
  that they appear as 'customers' directly when using filters
- ensure an order that fails to synchronizes is skipped and does not
  crash the whole cron (missing savepoint)
- correctly handle inactive currencies when syncing orders

See the initial PR for 13.0 for more details and a better split between
code changes.

Task-2279418

closes odoo/enterprise#12506

Related: odoo/upgrade#1657
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>
Co-authored-by: Antoine Vandevenne <anv@odoo.com>
Co-authored-by: Damien Bouvy <dbo@odoo.com>