[IMP] timesheet: add hours next to avatar widget
Implement a new widget extending the standalone avatar widget.
Its goal is to display the number of hours an employee have to work on
the currently displayed period (on the grid view). It's displayed in
red if all the hours have not been timesheeted AND are in the past (so
previous week or previous month). Otherwise, it's gray (muted style).

TASK ID 2264629

closes odoo/enterprise#12954

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>