[FIX] social_facebook: fix raise when accessing to a deleted post
When a post is deleted on Facebook, his social.post counterpart is still
available in the database.

This commit introduces a small fix to avoid a raise when such a post is opened
in the Feed view.
We now display a nice error message saying that the post has been deleted.

Task-2302570

X-original-commit 1dc02335ad29606e5a0b52ef0edea349aa191741

X-original-commit: 2743b1f20e059c0394bddb878da6e436d283c6d3