[FIX] stock_barcode: scan package in picking batch
Before this commit, when the user scans an unexpected package in a
picking batch, they get a traceback.
That's because we avoid to create new lines when scanning something in
batch picking, and `_step_package` don't take this restriction in count
and try to increment a line who doesn't exit.

How to reproduce:
  - Create a batch picking with some pickings, confirms it;
  - Create a package at the same location then those pickings and add
    product in it;
  - Open the batch picking in the barcode app and scan the package.

With this commit, we enable the Barcode App to create a new line when
processing batch pickings too.
In the case described above, scanning a package available at the current
location will create a new line for the product with the quantity
contained by the package. The original line qty will stay unchanged.

task-2323279