[FIX] stock_barcode: scan different lot name
Before this commit, in the barcode app, if the user scans a unexpected
lot in a picking batch, the line will be incremented and the line's lot
will be changed, but the lot name will not be updated in the interface.
Now, the lot name will be correclty updated.

task-2323279