[FIX] voip: Refuse incoming call
Before this commit:
In voip, if we try to refuse an incoming call, the call stops but
starts again one second after. If we refuse a second time this call,
the call stops. So to refuse an incoming call, we must click twice
on cancel button.

After this commit:
Click once on the cancel button is enough to refuse an incoming call.
Technically, when rejecting the call, we send status code 603
(‘Decline’) instead of 480 (‘Temporarily Unavailable’).
More details: https://tools.ietf.org/html/rfc3261#section-21.6.2

TaskID 2309518

closes odoo/enterprise#12974

X-original-commit: e93b65f20088cd3e80f8cdc906043ba0b066a9ac
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>
Signed-off-by: jbm-odoo <jbm-odoo@users.noreply.github.com>