[FIX] web_gantt: draggable not working after collapse and expand group
When you collapse and expand the group header in Gantt view, the expanded row
will not be droppable. like when you drag the pill from different row and try
drop into collapsed and expanded row, pill will not be dropped as row is not
droppable once it is collapsed and expanded.

With this commit, we have resolved this issue by resetting the row droppable
again in updateRow.

Task : 2283977

closes odoo/enterprise#13796

X-original-commit: 5e63fc6df14b9d60ae128fdd9e8035e05f457427
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>
Co-authored-by: Mohammed Shekha