[IMP]: documents: support the link to an existing record
Add the possiblity to link/unlink a doc to an existing record (any odoo object) with an action ('link.to.record').
This action will open a wizard to select the model and record to link.
When configuring the 'link.to.record' actions, it is possible to set a model to use.
Actions of the type 'link.to.record' are not displayed in split document mode.
Adds buttons in the inspector document to edit or remove the link between a record and a document.
If a document is linked to a record and the file is replaced, the record chatter will only display the last document.

Renaming values in create_model.

Task-2195019

closes odoo/enterprise#12680

Related: odoo/upgrade#1810
Related: odoo/odoo#56554
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>