[FIX] sign: issue of popover and text selection behaviour
Currently, when preparing document for sign, popover is opened and we move field
it stay opened and did't follow the field.
when try to select text of field, text selection behave strange.

So, in this commit popover is going to closed when move field and
when text selection is disabled.

closes odoo/enterprise#12568

Taskid: 2295990
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: Mitali Patel <mpa@odoo.com>