[IMP] website_twitter_wall: add "favorites" search support for twitter walls
The commit adds the support to search for the favorited tweets of a specific
account. Currently, it's only possible to use the regular twitter search that
does not allow it.

This is mainly a requirement for this year's online Odoo Experience where we
want to be able to search for the favorited tweets of the Odoo account.

Task ID 2325137

Side note: this is a naïve and minimalist implementation for the stable branch.
There are probably many ways to implement this better, especially on the
functional point of view, but it would quickly require more code / more fields.

closes odoo/enterprise#12621

X-original-commit: 288c49a18fc69fe14d01ec3acdf005f58536fd16
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>