[IMP] account_invoice_extract: SWIFT code from OCR
OCR now detects SWIFT code.
When a new account is created for a new partner, if the bank corresponding to the detected SWIFT code exists in the DB we link this bank to the account.
If the bank doesn't exist but the SWIFT code has been verified by iap-extract (verified_bic=True) create the bank in the client's DB.

closes odoo/enterprise#12794

X-original-commit: cd9b6f73119d4ff3178be6c89e7b67f4e60e455a
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Sébastien Mottet <smottet@users.noreply.github.com>