[FIX] social_twitter: use a constraint instead of overriding "create"
Purpose
=======
On #12551 , we fixed a bug by adding a warning message in the create of
the stream type. This can not be implemented with a constraint in stable
because badly-configured streams (stream of type "twitter_follow" without
"twitter_followed_account_id" e.g. might exist in the database).

So, we do it in the right way on master with a migration script to fix
the badly-configured records.

Task-2302570
Migration - odoo/upgrade/pull/1706

closes odoo/enterprise#12538

Related: odoo/upgrade#1706
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>