[FIX/IMP] payment_sepa_sdd: support users inside a company
The current code would not allow a partner part of a company* to use the
sdd payment acquirer. The main issue being that a mandate is supposed to
be delivered by the company* rathar tnat one of its employees.

This commit ensures that we always assign the mandate to the
`partner.commercial_partner_id` and that we properly manage the pament
tokens related to these mandates.

* read: partner with a parent and that is not of type company.

opw-2270613

closes odoo/enterprise#14295

X-original-commit: f3b41c0a52881635d31314a5718f1365afaa1d63
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>