[FIX] hr_contract_salary: Fix resume creation issue
Currently, The name field is readonly in resume menu and when you
want to save the record it shows the error 'Required fields: Code '.
name field is not computed and not getting name from anywhere so
it should be editable and code should be only required if value_type
in 'playslip' or 'contract'.

So in this commit improves this behaviour by making the name field
editable in resume and code field wil be required based on attribute.

closes odoo/enterprise#14099

Taskid: 2326538
X-original-commit: 2b93112023445ffa18bdd8f0e2a06d78a9202c4b
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>