[FIX] stock_barcode: update qty_done
When processing on mobile, via the barcode app, a transfer with product
tracked by serial numbers, the quantity done isn't always updated.

How to reproduce (NEED TO BE ON MOBILE MODE):
  - Create a product tracked by Serial Number and set a barcode for it;
  - Put at least two qty. in stock for this product;
  - Create a delivery for at least 2 of these product;
  - Open this delivery with the barcode app;
  - On the form view, scan the product's barcode:
    -> It opens the barcode lot lines wizard;
  - Scan one of the SN:
    -> It updates correctly the lot qty. done (1/x);
  - Scan an another SN:
    -> It doesn't update the lot qty. done (still 1/x), even if the lot
       line qty. done is correclty to one.

To fix that, this commit adds an onchange on `stock_barcode.lot`
`stock_barcode_lot_line_ids` to force the update of the `qty_done`.

opw-2276723

closes odoo/enterprise#12773

X-original-commit: 19416be79ee4559f88bb0414f42559a2e846577a
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>