[FIX] timesheet_grid: Fix traceback when filter timesheets from portal
Before this commit, while selecting the group by on project and
applying filter on the last year OR group by on employee showing
the traceback on timesheet.

When passing the domain from portal the domain is a set of tuples
and we can not assign the value to the tuples.

So this commit, Fix the issue and convert the tuple to list so
we can assign th value to list element.

closes odoo/enterprise#12831

Taskid: 2326440
X-original-commit: ee92a35ace8abe266359b93620b7a5e49d47b8b9
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>