[REF] social: manually create social.account's utm.medium
This commit changes the implementation of utm.medium creation to manually
create utm.medium every time a social.account is created.

They used to be created using a "inherits" on utm.medium, but it had the big
disadvantage of not using the "name_get" of the social.account to display the
utm.medium name. The "name_get" adds the name of the related social.media in
front of the name of the account (e.g: "[Twitter] Odoo").

When adding several social.accounts, as your accounts will often be named the
same way (e.g: "Odoo"), you would end up with several utm.medium with the exact
same name, making it impossible to view clear statistics.

Now, we manually create an utm.medium and set its name correctly so that it
contains the related social.media. This way, all created utm.medium linked to
social.accounts will be properly named and it will help read statistics.

LINKS

Task ID-2206496
ENT PR odoo/enterprise#20383
UPG PR odoo/upgrade#2763

Co-Authored-By: Aurélien Warnon <awa@odoo.com>
Co-Authored-By: Thibault Delavallée <tde@odoo.com>