[IMP] account_auto_transfer: use the new accounting test framework
Not using them causes some tests to fail on some of runbot's l10n builds, as the currency may restrict the number of decimal places below 2. With this change, the tests are not run on l10n build by default (as these functions are unlikely to be affected by l10n custom stuff), and need to be overridden specifying the chart template to use if we want that to change in specific locations.

X-original-commit: ea34d73560d89f4f3b26668c1259c0d0680099c6