[ADD] social_crm: create lead from social post/comment
To allow user to generate a lead in the CRM, a 'Generate Lead' option has been
added in the dropdown menu associated to each social comment and post.

This option is available when the user has the appropriate rights and when he
is not the author of the post/comment.

When clicked, the new option will open a conversion wizard that allows linking
the created lead to an existing customer, or to a create one.
We also try to find a matching existing customer using a name_search with the
name of the author of the post/comment we are converting.

Since the "description" field is now a Html field, we use a specific template
to get a nice rendering containing the various information we have:
- Author name
- Post/comment content
- Published date
- Link to the post/comment

Please note that for obvious privacy reasons, the social media platforms will
not provide us with people's email, phone or address information, we only have
the author name and the content of his post/comment.
For that reason, end users will probably often only convert to leads based on
existing clients, since otherwise they will not have any mean to contact the
created lead.

The generate lead feature has been implemented for:
- Facebook
- Twitter
- LinkedIn
- Instagram
- YouTube
But can be implemented for future media easily by following the same
architecture.

To enforce our lead creation logic, a few tests were added into the
social_test_full module.

In addition, we improved the name of the utm.source created for each
social.post by including the few first characters of the message.
It will ease the reading of statistics related to the posts.

LINKS

Task ID-2206496
ENT PR odoo/enterprise#20383
UPG PR odoo/upgrade#2763

Co-authored-by: Aurélien Warnon <awa@odoo.com>