[FIX] account_yodlee: ensure balance of provider is updated
Before this commit, and since Odoo 10.0, the balance of a Yodlee
account.online.provider was only written at the very first
synchronization, but never updated after that because of a wrong line of
code.

It was not a problem before saas-13.3 because this value was never used
elsewhere. However, since 6136b991e1a Yodlee is calling
online_sync_bank_statement with self.balance as last argument, so this
value is now used.

The issue is that new synchronizations are always created with the same
wrong closing balance, which is the closing balance of the very first
synchronization.

closes odoo/enterprise#12807

X-original-commit: c9fb4a21f09f625278c8d3049deb9da2d0c1493a
Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>
Signed-off-by: Paul Morelle <madprog@users.noreply.github.com>