[FIX] web_gantt: sample mode: stay or leave
Before this commit, a gantt view with a default group by defined on its
arch and sample data enabled would leave the sample mode when switching
to another view and going back. The same would happen when triggering a
reload by pressing enter while having the focus in the search bar.
This used to happen because the key 'groupedBy' in loadParams was
modified in place in __load, so that the function _haveParamsChanged
(called at reload) would return true while nothing would really change
(in the control panel or the gantt view).

The present commit fixes that situation.

Task ID: 2329559

closes odoo/enterprise#13101

X-original-commit: b5a138f657a7a6f43dd633a384af09e3f7147c01
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>