[FIX] helpdesk_fsm: Fix traceback while click on 'Plan intervention'
Currently, In the Helpdesk ticket when we click on 'Plan intervention'
button, it will raise an error.
when going to take the project_id using default_get of task at that
time it was return the dict of default_get values and in that there is
value for timesheet_product_id and which field does not exist in create
fsm task wizard.

So in this commit, We fix this traceback by passing the only project_id
value in default get of fsm create task wizard.

closes odoo/enterprise#13156

Taskid: 2328594
X-original-commit: 2c4ef60fdb87ad7755175abd0d282b0fb3e6536b
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>