[FIX] quality_control(_mrp_workorder): install qc without mrp
Usecase to reproduce:
- Start a new database and install quality_control module

Crash because finished_lot_id is in the view but not in the model.

finished_lot_id is defined in mrp_workorder. This fix move the
definition in the wrong view to mrp_workorder that define the field.

Also move lot_id since it's mainly used by mrp

closes odoo/enterprise#12932

Task: 2330107
X-original-commit: b8bf2f47bf60ceb524211ebd269a2f86697117f1
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>
Signed-off-by: Xavier Dollé (xdo) <xdo@odoo.com>