[IMP] documents: issue of remaining page lost when click on discard
Currently, When we click on `Discard` after a first action on the split tool,
the remaining pages disappears, there is no remaining document.

So, in this commit added archive button at top right to archive remaining pages
when click on archive button selected page from remaining pages are stored in
archive. and to come back to documents app add cancel button.

closes odoo/enterprise#12911

Taskid: 2277765
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>