[IMP] web_enterprise, sign: Hide empty cp-bottom on mobile
Currently, when o_cp_bottom is not filled, it will  take place(space)
in mobile.
It was going to take place due to min-height set on the o_cp_bottom.
also iframe has fixed height for sm device so it will show sapce at bottom

So in this commit, we set the min-height auto when there is mobile
view so it should not aquire the space when it's empty.
Fix the bottom height with 100% height in mobile instead of fix width.

closes odoo/enterprise#12922

Taskid: 2319534
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>