[FIX] web: avoid displaying HomeMenu when url state is changing
On a fresh db, in debug=0
go to settings, change the paper format
Save

Before this commit, the HomeMenu was briefly displayed before returning
back to the settings view.
This was because when saving the url state changes, triggering the
handler for hashchange multiple times. The DropPrevious handling
the loading of the menus rejected, which made the HomeMenu display

After this commit, the HomeMenu does not appear during a transition
from action to action.

taskid 2301277

closes odoo/enterprise#12934

X-original-commit: b5f10d33ca6b6d284bf557de0c8a4d9c084a22a4
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>