[IMP] mrp_workorder: workorder step if operation
Quality point's `is_workorder_step` is true if it has already an
operation set as operations are MRP depending.
The purpose is to display the operation field when creating a new
operation step (set by default when coming from the BOM form view, but
invisible because the operation Types field is empty).

task-2328830

X-original-commit: f68ed08df8a028d8a77345243449c12f54f03edc