[FIX] accountant: reconciliation widget: make write-off rules work with partner mapping
Before that, write-off rules with partner mapping did not work properly, as the partner they selected got lost when passing the values to the widget, which later made a second rpc call to fetch the write-off details. Now, we pass all the values directly, including the selected partner. This way, we remove the need for a second rpc call when processing the results for the reconcile models, and the partner is handled properly.

closes odoo/enterprise#13083

X-original-commit: 79595796febb23e01eac34f6842b4fe88cfb4b6b
Related: odoo/odoo#57358
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>