[FIX] sign: signature out of the box
Before this commit:

When the user uses initial or signature fields to put the signature into
the document, some parts of the signature been seen outside the box after
document validation.

After this commit:

The signature will be inside the box at the correct place by flex display.

LINKS

Task-2328677

closes odoo/enterprise#13178

X-original-commit: 7b1398673ae561561292624f4326f09050353853
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>