[IMP] currency_rate_live: Chilean currency update
Include mindicador.cl connection, to update
the following currencies: USD and EUR. Also
the connection can update to financial indexes: UF
(stands for "Unidad de Fomento") and UTM (stands
for "Unidad Tributaria Mensual").
These two currencies (USD and EUR) are the most
common currencies used in Chile for international
transactions and the rates supplied by the site "mindicador.cl".
The UF is an index that is forecasted in a daily
basis using the IPC (consumer prices index) and
is an intensively used value in different business
transactions, to keep the amounts independent from
inflation.
The UTM is an index used for taxes or fines
adjustments or money returns used by the
government. Even when this value is published on
a monthly basis.

X-original-commit: e24ac3a59d43c356e9d8d9f70441b40e0b71ef5e