[FIX] hr_work_entry_contract: Improve performances on work entries generation
Purpose
=======

When generating the work entries the first time for older contracts,
we try to generate the work entries for several months/years, instead of
only the current month for example.

The reason if that the date_generated_from and date_generated_to fields
are there to avoid having gap on the generated work entries.

If they were generated from January to March, and we try to generate
them in July, then we generate the work entries for April, Mai, June,
and July, so that if we remove work entries on a given day, we don't
try to re-generate them afterward.

If we move the cursors to the start date when they are equal, instead of
generating 138300 records for 100 employees who have a contract starting
on the 1rst of January 2018, to generate work entries till the 31nd of
August 2020, then generate 7900 records.

This represents 17.5 times less records to create, and it allows to use
generate the records on real database for the first time without risking
to face a timeout.

Taskid: 2321853
X-original-commit: 5b85ec0a2daa6df731a61f332b71a48a71015d06