[FIX] hr_work_entry_contract: Avoid creating empty work entries
Purpose
=======

It could happen that some hr.leave are created at a period during the
employee is not supposed to work.

This results into empty work entries, for which the start and end
dates are equal, leading to a sql constraint error.

Avoid generating empty work entries for that use case.

closes odoo/enterprise#13085

Taskid: 2321853
X-original-commit: 0303d6b96a6e6f7ee501d351d0417b280064cde6
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>