[IMP] documents(_spreadsheet): Disable the buttons in readonly workspace
Purpose of the task is, When a user has only read access to a workspace
The top buttons(upload/add url/request/... etc) and side panel buttons
should be disabled.

So in this commit, If the read/write groups are defined on the workspace
(if not defined then default all user's have a write access to that
workspace) then disable the control panel buttons and document's sidebar
buttons based on workspace rights.
Also prompt Aceess error message on move of documents from one
workspace to another but when move using the action then we should allow
to movie the documents.

task-2086557

closes odoo/enterprise#12351

Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>