[FIX] Helpdesk: Fix error when enabling Time Reinvoicing
When deleting all Tickets & enabling Time Reinvoicing, an error occurs while loading data files
(more precisely records id helpdesk_timesheet_1 & helpdesk_timesheet_2).
That result on a reverse of this data file.
While being reversed, odoo still update data in database and makes a link between a record and an unexisting project.
Because of the link with an unexisting project, an error occurs and the module failed to install.

This commit removes the noupdate="1" so that there is no error when activating Time Reinvoicing

Task id 2326300

X-original-commit: 1cdf402d090c7410080b8f7fbc98a7fdbcf71db6