[IMP] l10n_be_hr_payroll: Allow to export/import commissions before generation
Purpose
=======

Commissions are paid every 3 months. Amount of commission is specific to sales
so we need a wizard to import easily the amount for each sales.

The flow: the CFO provides a list of sales with an amount, and the hr import it in the tool,
the system computes the net (precompte and onss are computed like classical salary).

closes odoo/enterprise#13132

Taskid: 2300044
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>