[FIX] planning: make unchecking recurrence work
Unchecking the repeat field on the form of a recurrent slot didn't work.
Thanks to adt for finding the problem, and the fix applied in this commit.

His explanation:
"The reason for which it didn't work: all the fields `repeat_*` are protected during computation of the inverse, and the only field that is cached is `repeat` because it has just been set to false, on the other hand all the other fields are not cached and can't be computed either (since they're protected), so the default behavior when a value is not cached and it can't be computed is to return `False`"

Also there were some discrepancies between comment/code in the tests. Some test were also
straight up not testing what they claimed to be testing. All of this
made expected behavior hard to understand.

task-2326229

closes odoo/enterprise#14991

X-original-commit: 51985c19396107b71d9aa681abb9dead2a0a3d3b
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>