[FIX] Helpdesk: No permission when clicking on button on main page
Before this commit, when clicking on a button on the main page (My tickets, Avg Open Hours, ...), the user got a permission error
This occures because the user didn't have admin permission.

This commit opens the view in sudo mode so the user don't have the permission error.

Task id: 2326300

closes odoo/enterprise#13842

X-original-commit: fff54ca014a53039ef3b7ade68c89b97a60c59ac
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>