[FIX] mrp_workorder: BOM operations steps
Before this commit, when creating a new operation step when coming from
the BOM form view, if the user go back to the operation steps list via
the breadcrumbs, the newly created step doesn't appear in the list.
That was due to the domain passed to the action who filter on fixed
step ids, and so, doesn't take all BOM's operation steps dynamically.

task-2328830

closes odoo/enterprise#13876

X-original-commit: 2e5434cb145364cd998e9ec88c68cf8ada4652c2
Related: odoo/odoo#59398
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>