[FIX] web_map: fetch coords asynchronously
Before this commit, coordinates were fetched all in once
because of that, we got banned from openstreetmap's server every time
we load the view, it took a long time to load.

Now, model doesn't wait the coordinates to be fetched anymore
but still fetch them and trigger an update of the view
when a coordinate is fetched. Coordinates are fetched every seconds
to not get banned from openstreetmap's server.

Task id: 2320271

closes odoo/enterprise#13807

X-original-commit: 656cb56a17c2d9d2168763cb5679a575a11b6113
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Michaël Mattiello <mcm-odoo@users.noreply.github.com>