[IMP] industry_fsm: Adapt planned date to company/resource calendar
**Purpose**
Currently, when creating a new task from the Gantt view, the
planned dates set as default are Monday 00:00:00 - Sunday 23:59:59.
This is unlikely to happen in a real-life scenario.
Adapting to the hours defined on the company/resource calendars
would greatly ease the encoding.

**Specifications**
-when creating a task from the gantt view, set today's date by
default;
-if the task is unassigned, the default start/end hours should
correspond to what is configured on the company calendar for that
day of the week;
-If the task is assigned to a specific user, the default start/end
hours should correspond to what is configured on the resource
calendar of the employee for that day of the week.

Task ID 2318711
Enterprise PR #13194

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>