[IMP] documents_spreadsheet: add missing pivot cells
If a pivot has values that evolve over time, the user needs a way to
quickly find the missing cells or to be able to quickly add any values
from a pivot to insert it in a given cell.

For that, we created a new dialog to display the pivot and to allow the
user to select a cell. The missing values are displayed with a yellow
background to easily find them.

To identify missing values, we simply mark the values used when evaluating
the pivot formulas.

Task-ID #2335931

closes odoo/enterprise#13310

X-original-commit: 75142a9ee6cf430f5189275d80897f45130128e2
Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Co-authored-by: fleodoo <fle@odoo.com>
Co-authored-by: Vincent Schippefilt (vsc) <vsc@odoo.com>
Co-authored-by: Pierre Rousseau <pro@odoo.com>