[FIX] mps: performance improvements
Before in a DB with 6000 `mrp.production.shedule`, 6800 product and
3000 bom the loading of the "Master Production Schedule" was
very slow:
- For 20 `mrp.production.shedule`: 19.63 +- 0.93 sec
- For 50 `mrp.production.shedule`: 71.94 +- 0.45 sec

To improve performance:
- Avoid filtering in a loop, use dict + one loop.
- Refactor `_get_indirect_demand_ratio` to be batched
(`_get_indirect_demand_ratio_mps`) and tranverse
only one time the tree.
- Cached at the most of possible results of mapped.
- Avoid `in` check in list (in a loop) and use set or
OrderedDict instead.

Also the hidden bom was fixed. Before
if a product doesn't have a `mrp.production.schedule`
but was in the tree of production of a other product,
it wasn't take in account. Fallback
to bom_find in this case.
(Unfortunately it descreases performance).

After this commit, performance is far better but still can be improve:
- For 20 `mrp.production.shedule`: 2.8 +- 0.13 sec
- For 50 `mrp.production.shedule`: 4.07 +- 0.26 sec

opw-2085309

closes odoo/enterprise#13330

X-original-commit: 8176a0a198a16d59954c73e3f307c5ca16c2a5c9
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>